Refactor feature parsing
Features are now parsed directly from the TokenStream using the Parse
trait, which allows for parsing without multiple nested `match`
statements. Instead, we parse directly into a structure that matches the
tokens of our feature grammar, which can then be directly turned into
a feature.

In addition, features are now parsed in a pipeline fashion, which allows
for cleaner separation of successful vs unsuccessful parses.